refactor: extract color palettes from main.rs
