+++ src/tool.rs
@@ -1,4 +1,5 @@
 //! This module contains interactive tools, like
 //! color palette generators.
 
-pub mod color;
+mod color;
+pub mod palette;
